bug fixes in viewSnowShoeTour

+++ src/main/java/ca/mcgill/ecse/snowshoetours/controller/GearController.java
@@ -121,6 +121,8 @@ public class GearController {
     for(Combo combo : combos) {
       if(combo.getName().equals(name)) aCombo = combo;
     }
+    if(aCombo == null) return;
+    
     aCombo.delete();
   }
